# docs/swagger.yaml
definitions:
  handlers.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  handlers.SuccessResponse:
    properties:
      data:
        description: Опциональные данные
      message:
        type: string
    type: object
...
  models.Group:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      student_count:
        type: integer
      updated_at:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /api/group:
    get:
      consumes:
      - application/json
      description: Возвращает список всех групп
      produces:
      - application/json
      responses:
        "200":
          description: 'Успешный ответ"  example([{"id": 1, "name": "Group A"}, {"id":
            2, "name": "Group B"}])'
          schema:
            items:
              $ref: '#/definitions/models.Group'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить список всех групп
      tags:
      - Groups
    post:
      consumes:
      - application/json
      description: Создаёт новую группу
      parameters:
      - description: Данные о группе
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Создать группу
      tags:
      - Groups
  /api/group/{id}:
    delete:
      consumes:
      - application/json
      description: Удаляет группу по её ID
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Удалить группу
      tags:
      - Groups
    get:
      consumes:
      - application/json
      description: Возвращает информацию о конкретной группе
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/models.Group'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Группа не найдена
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Получить информацию о группе по её ID
      tags:
      - Groups
    put:
      consumes:
      - application/json
      description: Обновляет информацию о существующей группе
      parameters:
      - description: ID группы
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: Обновлённые данные о группе
        in: body
        name: group
        required: true
        schema:
          $ref: '#/definitions/models.Group'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/handlers.SuccessResponse'
        "400":
          description: Неверный запрос
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      summary: Обновить информацию о группе
      tags:
      - Groups
...
swagger: "2.0"